+++ src/Navbar.js
@@ -1,33 +1,43 @@
 import React, { useRef, useState } from 'react'
-import BigMenu from './BigMenu';
-import { ReactComponent as BigLogotype} from './Images/agencylogotype.svg';
+import BigMenu from './BigMenu'
+import { ReactComponent as BigLogotype } from './Images/agencylogotype.svg'
 
 function Navbar(props) {
   const [bigMenu, determineBigMenu] = useState('no')
-  const showBigMenu = (() => {
+  const showBigMenu = () => {
     if (bigMenu === 'no') {
       determineBigMenu('yes')
     }
-    if (bigMenu === 'yes') {determineBigMenu('no')}
-  })
+    if (bigMenu === 'yes') {
+      determineBigMenu('no')
+    }
+  }
   const hamburgerMenuRef = useRef()
 
   return (
     <>
-      {bigMenu === 'yes' && <BigMenu/>}
+      {bigMenu === 'yes' && <BigMenu />}
       <div className={props.passClass}>
-        <div><BigLogotype className='navbar-logotype' /></div>
-        <div className='navbar-text'>
-            <div className="">Home</div>
-            <div className="">About</div>
-            <div className="">Services</div>
-            <div className="">Contact</div>
+        <div>
+          <BigLogotype className="navbar-logotype" />
         </div>
-        <div ref={hamburgerMenuRef} className='hamburger-menu' onClick={() => {showBigMenu()}}>
+        <div className="navbar-text">
+          <div className="">Home</div>
+          <div className="">About</div>
+          <div className="">Services</div>
+          <div className="">Contact</div>
+        </div>
+        <div
+          ref={hamburgerMenuRef}
+          className="hamburger-menu"
+          onClick={() => {
+            showBigMenu()
+          }}
+        >
           {bigMenu === 'no' && <span>≡</span>}
           {bigMenu === 'yes' && <span>×</span>}
         </div>
-    </div>
+      </div>
     </>
   )
 }
