+++ src/Pages/home/Components/appeal.jsx
@@ -1,0 +1,20 @@
+import React from 'react'
+
+function Appeal() {
+  return (
+    <div>
+      <h1 className="centered">Any other way we can help?</h1>
+      <p className="centered">
+        Lorem ipsum dolor sit amet consectetur adipisicing elit. Cumque
+        excepturi facilis accusantium odit dicta? Aperiam soluta veritatis
+        ratione? Reiciendis officia ipsam esse recusandae, cupiditate maiores
+        deserunt autem, quod quisquam vitae inventore voluptas porro optio
+        commodi veritatis voluptates similique. Earum perferendis maiores
+        pariatur consectetur nostrum tempore aspernatur autem, cumque iste
+        corporis.
+      </p>
+    </div>
+  )
+}
+
+export default Appeal
