--- src/components/index.js
@@ -1,2 +1,0 @@
-export { default as Button } from './Button'
-export { default as Logo } from './Logo'
