+++ src/index.js
@@ -2,7 +2,6 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import './index.css'
 import App from './App'
-import '@fontsource/khula'
 import { createRoot } from 'react-dom/client'
 import GlobalStyles from './styles/GlobalStyles'
